winevulkan: Refactor fs hack to allow for multiple upscaler backends
